+++ packages/create-theme/public/template/docs/.vitepress/theme/index.ts
@@ -1,4 +1,4 @@
-import BlogTheme from '@sugarat/theme'
+import BlogTheme from '@desain/theme'
 // 默认颜色仍然使用绿色
 import './green-theme.var.css'
 // 自定义样式重载