+++ packages/theme/docs/changelog.md
@@ -12,7 +12,7 @@ publish: true
 升级主题，原项目执行如下指令即可
 
 ```sh
-pnpm up @sugarat/theme
+pnpm up @desain/theme
 ```
 
 ## 0.1.41 (2023/08/13)
@@ -65,7 +65,7 @@ pnpm up @sugarat/theme
 
 ### Patch Changes
 
-- feat: 新的项目创建方式使用 [@sugarat/create-theme](https://github.com/Desain7/personal-blog/tree/master/packages/create-theme) 快速创建模板项目
+- feat: 新的项目创建方式使用 [@desain/create-theme](https://github.com/Desain7/personal-blog/tree/master/packages/create-theme) 快速创建模板项目
 - feat: 支持使用自定义的HTML内容设置文章在首页卡片列表里展示的 `描述信息`
 - feat: 支持使用 `cover:false` 隐藏首页展示的封面